refactor(default-tag): remove tag logic 변경하면서 필요 없는 부분을 전부 들어냈습니다.
